swype: character tracking using okboard's libcurve
Track keyboard swype using libcurve calls hooked into the touchpoints
in the PressArea.qml

In order to make this work, the downward swyping to hide keyboard is
disabled.  Instead, can hold the symbol key down for about 4 seconds to
hide the keyboard.  I doubt this is what we want in the end, but it works
and is easy.